I forgot to commit regularly so this is my first Commit
I work on the game syntax and built the program to compare the human Choice and Computer Choice and determine which one wins

I set up a better interface for the game, a score tracker, Buttons to choose an option.

I also set up Game effects so when the user/computer choose an option it shows on the screen.
So the game is more Interactive.

I also added an "Opening Scenario" so when the user first comes to the game screen it triggers a reaction. I used Interaction Observe API to achieve this.